txt file datasets

+++ CustomData/customdata_v2.py
@@ -15,18 +15,15 @@ from augmentation import augment_img
 
 class RoiDataset(Dataset):
 
-    def __init__(self, filename, train=True):
+    def __init__(self, root_dir, filename, train=True):
         """
         函数构造函数是imdb, imdb 是类 pascal_voc, 把该类作为初始化参数加到构造函数里面
         :param imdb: 类名pascal_voc
         :param train:
         """
         super(RoiDataset, self).__init__()
-        self._year = "2012"
-        self._image_set = "train"
-        self._devkit_path = "/home/chenxi/dataset/VOCdevkit"
-        self.data_dir = "/home/chenxi/dataset/VOCdevkit"
-        self.name = 'voc_' + self._year + '_' + self._image_set
+        self.root_dir = root_dir
+        self.train = train
         self.totalData = []
         self.imageName = []
         self.boxes = []  # boxes  [ [box], [[x1,y1,x2,y2], ...], ... ]
@@ -34,9 +31,8 @@ class RoiDataset(Dataset):
         self.mean = (123, 117, 104)  # RGB 形式的均值
         self.num_samples = 0  # 样本总数
 
-        self._data_path = "/home/chenxi/dataset/VOCdevkit/VOC2012/"
+        self.image_dir = "/home/chenxi/dataset/VOCdevkit/VOC2012/JPEGImages"
 
-        # self.num_classes 是父类的方法, 在此处继承父类方法,且父类中将其变成了属性,所以直接调用属性即可.
         self._class_to_ind = dict(zip(cfg.classes, range(len(cfg.classes))))  # 类别对应数字序号
 
         self.config = {'cleanup': False,
@@ -46,14 +42,8 @@ class RoiDataset(Dataset):
                        'rpn_file': None,
                        'min_size': 2}
 
-        # list, 存放图片的路径
-        image_set_file = os.path.join(self._data_path, 'ImageSets', 'Main', self._image_set + '.txt')
-        assert os.path.exists(image_set_file), 'Path does not exist: {}'.format(image_set_file)
-        with open(image_set_file) as f:
-            image_index = [x.strip() for x in f.readlines()]
-        self.image_index = image_index
         # ==================================
-        with open("/home/chenxi/tempfile/YOLO_v1/utils/voc2007test.txt") as f:
+        with open(filename) as f:
             lines = f.readlines()
         for line in lines:
             splited = line.strip().split()  # ['005246.jpg', '84', '48', '493', '387', '2'] img_name + 坐标 + 类型(labels)
@@ -83,21 +73,19 @@ class RoiDataset(Dataset):
         # _roidb 是一个list, 存放的是每一个xml内部标签,写成字典的格式{"boxes":array([[x1, y1, x2, y2]], "gt_classes":array([[label]]}
         self._roidb = self.load_data()
 
-        self.train = train
-
         self._image_paths = [self.image_path_from_index(self.imageName[i]) for i in range(len(self.imageName))]
 
     def image_path_from_index(self, index):
         """
         Construct an image path from the image's "index" identifier.
         """
-        image_path = os.path.join(self._data_path, "JPEGImages", index)
+        image_path = os.path.join(self.image_dir, index)
         assert os.path.exists(image_path), 'Path does not exist: {}'.format(image_path)
         return image_path
 
     @property
     def cache_path(self):
-        cache_path = os.path.abspath(os.path.join(self.data_dir, 'cache'))
+        cache_path = os.path.abspath(os.path.join(self.root_dir, 'cache'))
         if not os.path.exists(cache_path):
             os.makedirs(cache_path)
         return cache_path
@@ -109,8 +97,8 @@ class RoiDataset(Dataset):
         im_data = Image.open(image_path)
         boxes = self.boxes[i]
         gt_classes = self.labels[i]
-        print(self.totalData[i]['boxes'])
-        print(boxes)
+        # print(self.totalData[i]['boxes'])
+        # print(boxes)
 
         # 获得原始图像的 w, h
         im_info = torch.FloatTensor([im_data.size[0], im_data.size[1]])
@@ -154,59 +142,20 @@ class RoiDataset(Dataset):
         Return the database of ground-truth regions of interest.
         This function loads/saves from/to a cache file to speed up future calls.
         """
-        cache_file = os.path.join(self.data_dir, self.name + '_gt_roidb.pkl')
-        # if os.path.exists(cache_file):
-        #     with open(cache_file, 'rb') as fid:
-        #         roidb = pickle.load(fid)
-        #     print('{} gt roidb loaded from {}'.format(self.name, cache_file))
-        #     return roidb
+        cache_file = os.path.join(self.root_dir, self.cache_path, 'train_gt_roidb.pkl')
+        if os.path.exists(cache_file):
+            with open(cache_file, 'rb') as fid:
+                roidb = pickle.load(fid)
+            print('gt roidb loaded from {}'.format(cache_file))
+            return roidb
 
-        gt_roidb = [self._load_pascal_annotation(index) for index in self.image_index]
         with open(cache_file, 'wb') as fid:
-            pickle.dump(gt_roidb, fid, pickle.HIGHEST_PROTOCOL)
+            pickle.dump(self.totalData, fid, pickle.HIGHEST_PROTOCOL)
         print('wrote gt roidb to {}'.format(cache_file))
-        return gt_roidb
-
-    def _load_pascal_annotation(self, index):
-        """
-        Load image and bounding boxes info from XML file in the PASCAL VOC
-        format.
-        """
-        filename = os.path.join(self._data_path, 'Annotations', index + '.xml')
-        tree = ET.parse(filename)
-        objs = tree.findall('object')
-
-        # Exclude the samples labeled as difficult
-        if not self.config['use_diff']:
-            non_diff_objs = [obj for obj in objs if int(obj.find('difficult').text) == 0]
-            objs = non_diff_objs
-        num_objs = len(objs)
-
-        boxes = np.zeros((num_objs, 4), dtype=np.uint16)
-        gt_classes = np.zeros((num_objs), dtype=np.int32)
-
-        # Load object bounding boxes into a data frame.
-        for ix, obj in enumerate(objs):
-            bbox = obj.find('bndbox')
-            # Make pixel indexes 0-based
-            x1 = float(bbox.find('xmin').text) - 1
-            y1 = float(bbox.find('ymin').text) - 1
-            x2 = float(bbox.find('xmax').text) - 1
-            y2 = float(bbox.find('ymax').text) - 1
-
-            cls = self._class_to_ind[obj.find('name').text.lower().strip()]
-            boxes[ix, :] = [x1, y1, x2, y2]
-            gt_classes[ix] = cls
-
-        return {'boxes': boxes, 'gt_classes': gt_classes, }
+        return self.totalData
 
     def __len__(self):
         return len(self._roidb)
-
-    def __add__(self, other):
-        self._roidb = self._roidb + other._roidb
-        self._image_paths = self._image_paths + other._image_paths
-        return self
 
 
 def detection_collate(batch):
@@ -248,7 +197,7 @@ def detection_collate(batch):
 
 
 if __name__ == "__main__":
-    data = RoiDataset("voc_2012_train")
+    data = RoiDataset("/home/chenxi/dataset/VOCdevkit", "/home/chenxi/tempfile/YOLO_v1/utils/voc2007test.txt")
     i = 0
     print(data[i].__len__())
     print(data[i][0].shape)
